backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio==5.11.3
aiosqlite==0.20.0
aiofiles==23.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
apscheduler==3.10.4
firebase-admin==6.5.0
Pillow==10.4.0
